backend/requirements.txt
# ─── Django core ──────────────────────────────────────────────────────────────
Django==4.2.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2

# ─── Auth ─────────────────────────────────────────────────────────────────────
djangorestframework-simplejwt==5.3.1

# ─── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
django-encrypted-model-fields==0.6.5

# ─── Celery + scheduler ───────────────────────────────────────────────────────
celery==5.3.6
redis==5.0.4
django-celery-beat==2.6.0
django-celery-results==2.5.1

# ─── Vector DB ────────────────────────────────────────────────────────────────
qdrant-client==1.9.1

# ─── AI / ML ──────────────────────────────────────────────────────────────────
openai==1.30.5
anthropic==0.28.0
langchain-text-splitters==0.2.0
sentence-transformers==2.7.0

# ─── Source connectors ────────────────────────────────────────────────────────
slack-sdk==3.27.2
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
notion-client==2.2.1

# ─── Document processing / OCR ────────────────────────────────────────────────
pytesseract==0.3.10
pdfplumber==0.11.0
python-docx==1.1.2
Pillow==10.3.0
beautifulsoup4==4.12.3

# ─── Utilities ────────────────────────────────────────────────────────────────
requests==2.32.2
httpx==0.27.0
python-dateutil==2.9.0
cryptography==42.0.7
pydantic==2.7.3
